Added a bunch of Monogame level 3 code! + new tileset.
This commit added a number of resources I developed for my monogame homeworks, which should serve as a good start to the game. This includes a basic sprite and animation system, a player character that can be controlled and is animated, tiled integration (this one with a new tilemap for the project, rather than the one that I used for the homework), and a few other features, like boxes, pickups, and a goal object. This should give us a big head start on the project. We'll still need to edit it to better fit our vision for the game, but from here out it should be much easier to parallelize work now that we have a base to work off of.